feat(fiori-elements-writer): separate fpm logic
